change solution using DateTimeFormatter and Duration feats

+++ src/main/java/edu/hw5/task1/TimeHoursMinutes.java
@@ -1,5 +1,5 @@
 package edu.hw5.task1;
 
-public record TimeHoursMinutes(int hours, int minutes) {
+public record TimeHoursMinutes(long hours, long minutes) {
 
 }
